requirements.txt
azure.core
azure-ai-documentntelligence
streamlit
azure-storage-blob
python-dotenv